api-service/requirements.txt
# Core FastAPI dependencies
fastapi==0.128.7
uvicorn[standard]==0.34.0
gunicorn==23.0.0

# Database dependencies
sqlalchemy==2.0.46
alembic==1.14.1
psycopg2-binary==2.9.11
asyncpg==0.31.0

# Authentication and security
joserfc==1.0.1
pwdlib[bcrypt]==0.2.1
bcrypt==4.3.0
python-multipart==0.0.20
cryptography==46.0.0

# Validation and serialization
pydantic==2.10.6
pydantic-settings==2.8.1

# HTTP client and utilities
httpx==0.28.1
aiofiles==24.1.0

# Protocol clients for connection testing
paho-mqtt==2.1.0
confluent-kafka==2.13.0

# Caching and session management
redis==5.2.1

# Monitoring and logging
structlog==25.1.0
prometheus-client==0.21.1

# Development utilities
python-dotenv==1.0.1
psutil==6.1.1

# CSV processing
pandas==2.2.3
numpy==2.1.3

# Background tasks
celery==5.4.0

# Rate limiting
slowapi==0.1.9

